+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v3.19.4
+#### **resource-metadata-sqs**
+### 🐛 Bug Fix 🔧
+- Add missing env var `EC2_CHUNK_SIZE`
+- Add inline policy to assume cross-account AWS Config role
+
 ## v3.19.3
 #### **resource-metadata-sqs**
 ### 🐛 Bug Fix 🔧